Add multitile Tile support / Make located tile usable

+++ PokeSi/DrawHelper.cs
@@ -28,8 +28,8 @@ namespace PokeSi
 
         public static Rectangle ExtendToContain(Rectangle toExtend, Rectangle toContain, int centerX = 0, int centerY = 0)
         {
-            float horizontalRatio = toContain.Width / toExtend.Width;
-            float verticalRatio = toContain.Height / toExtend.Height;
+            float horizontalRatio = toContain.Width / (float)toExtend.Width;
+            float verticalRatio = toContain.Height / (float)toExtend.Height;
             float ratio = Math.Max(horizontalRatio, verticalRatio);
             Rectangle result = toExtend;
             result.X -= (int)(centerX * (ratio - 1));
